model/requirements.txt
absl-py @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e3w1nvit39/croot/absl-py_1714140475537/work
aiodns @ file:///opt/miniconda3/conda-bld/aiodns_1762184801219/work
aiohappyeyeballs @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_86beu757ao/croot/aiohappyeyeballs_1755769847762/work
aiohttp @ file:///opt/miniconda3/conda-bld/aiohttp_1762519752188/work
aiosignal @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e3cjtbh902/croot/aiosignal_1755870814211/work
anyio @ file:///opt/miniconda3/conda-bld/anyio_1758623411472/work
async-timeout @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_0fagvkecl5/croot/async-timeout_1732662297241/work
attrs @ file:///opt/miniconda3/conda-bld/attrs_1762356889054/work
Bottleneck @ file:///opt/miniconda3/conda-bld/bottleneck_1761937849374/work
brotlicffi @ file:///opt/miniconda3/conda-bld/brotlicffi_1764961161728/work
certifi @ file:///opt/miniconda3/conda-bld/certifi_1763135327670/work/certifi
cffi @ file:///opt/miniconda3/conda-bld/cffi_1761832772717/work
charset-normalizer @ file:///opt/miniconda3/conda-bld/charset-normalizer_1761744841364/work
datasets @ file:///opt/miniconda3/conda-bld/datasets_1764933294060/work
dill @ file:///opt/miniconda3/conda-bld/dill_1760096784226/work
exceptiongroup @ file:///opt/miniconda3/conda-bld/exceptiongroup_1761560357548/work
filelock @ file:///opt/miniconda3/conda-bld/filelock_1762419675764/work
frozenlist @ file:///opt/miniconda3/conda-bld/frozenlist_1761062017275/work
fsspec @ file:///opt/miniconda3/conda-bld/fsspec_1762432466138/work
gmpy2 @ file:///opt/miniconda3/conda-bld/gmpy2_1764747893574/work
grpcio @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_cfvg1gbxnd/croot/grpc-split_1742490251423/work
h11 @ file:///opt/miniconda3/conda-bld/h11_1761931258948/work
hf-xet @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_79w069vczw/croot/hf-xet_1755725955657/work
httpcore @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_941yf7p5zt/croot/httpcore_1748526061031/work
httpx @ file:///opt/miniconda3/conda-bld/httpx_1760447189111/work
huggingface_hub @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_29dc5yixdo/croot/huggingface_hub_1755770856782/work
idna @ file:///opt/miniconda3/conda-bld/idna_1761912002779/work
importlib_metadata @ file:///opt/miniconda3/conda-bld/importlib_metadata-suite_1763996574720/work
Jinja2 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_2cnn4kenrm/croot/jinja2_1741710859444/work
Markdown @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_c8yxkefe2v/croot/markdown_1746114450220/work
MarkupSafe @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_1f_uj4vxik/croot/markupsafe_1738584045311/work
mpmath @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_17iu6a8a3m/croot/mpmath_1690848269369/work
multidict @ file:///opt/miniconda3/conda-bld/multidict_1760566705389/work
multiprocess @ file:///Users/ec2-user/croot/multiprocess_1766144188358/work
networkx @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_e6lt1dcyid/croot/networkx_1737040482044/work
numexpr @ file:///opt/miniconda3/conda-bld/numexpr_1762165517928/work
numpy @ file:///opt/miniconda3/conda-bld/numpy_and_numpy_base_1761579345363/work/dist/numpy-2.2.5-cp310-cp310-macosx_12_0_arm64.whl#sha256=cee7bef98ac83b78786789b80f44469231df1e3cf12ce327f909aa4ea0fcee88
opentelemetry-api @ file:///opt/miniconda3/conda-bld/opentelemetry-api_1763996076035/work
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079158/work
pandas @ file:///opt/miniconda3/conda-bld/pandas_1762332182781/work/dist/pandas-2.3.3-cp310-cp310-macosx_12_0_arm64.whl#sha256=15e77fd5caae22506c552ef9e15f2690708d3395500d8e3ac8ce5c5cf227c7a1
pillow @ file:///opt/miniconda3/conda-bld/pillow_1762528235859/work
propcache @ file:///opt/miniconda3/conda-bld/propcache_1765457305371/work
protobuf @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_73kluigbu7/croot/protobuf_1742426204819/work/bazel-bin/python/dist/protobuf-5.29.3-cp310-abi3-macosx_11_0_arm64.whl#sha256=05aeb246182cf9c9c5fff927f1c3af57f48c71a22404030bdf2f62e5dd8c5fdf
pyarrow @ file:///opt/miniconda3/conda-bld/pyarrow_1764939446398/work/python
pycares @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b01dqj7707/croot/pycares_1756217969873/work
pycparser @ file:///opt/miniconda3/conda-bld/pycparser_1757496094255/work
PySocks @ file:///opt/miniconda3/conda-bld/pysocks_1761753011241/work
python-dateutil @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_66ud1l42_h/croot/python-dateutil_1716495741162/work
pytz @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_3bjczh6it0/croot/pytz_1752135865541/work
PyYAML @ file:///opt/miniconda3/conda-bld/pyyaml_1763465492983/work
regex @ file:///opt/miniconda3/conda-bld/regex_1765378750161/work
requests @ file:///opt/miniconda3/conda-bld/requests_1762359594551/work
safetensors @ file:///opt/miniconda3/conda-bld/safetensors_1762515994044/work
six @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_03myqm7p6o/croot/six_1744271511946/work
sniffio @ file:///opt/miniconda3/conda-bld/sniffio_1764329153177/work
sympy @ file:///opt/miniconda3/conda-bld/sympy_1764763032713/work
tensorboard @ file:///opt/miniconda3/conda-bld/tensorboard_1764970467117/work/tensorboard-2.20.0-py3-none-any.whl#sha256=9dc9f978cb84c0723acf9a345d96c184f0293d18f166bb8d59ee098e6cfaaba6
tensorboard_data_server @ file:///var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_32p363irxt/croot/tensorboard-data-server_1724172561901/work/tensorboard_data_server-0.7.0-py3-none-macosx_11_0_arm64.whl#sha256=a3b45e50890dd6f2cb671ed443e84f614bee4eddfd0a51054a2d581cf02c78df
tokenizers @ file:///opt/miniconda3/conda-bld/tokenizers_1761922216898/work
torch==2.5.1
torchvision==0.20.1
tqdm @ file:///opt/miniconda3/conda-bld/tqdm_1762863350713/work
transformers @ file:///opt/miniconda3/conda-bld/transformers_1763719386871/work
typing_extensions @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_dffmcwqo6o/croot/typing_extensions_1756281471323/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3 @ file:///opt/miniconda3/conda-bld/urllib3_1765399198238/work
uvloop @ file:///opt/miniconda3/conda-bld/uvloop_1762175501861/work
Werkzeug @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_87am5fxnsf/croot/werkzeug_1737448721856/work
xxhash @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_4dqzb3ngyc/croot/python-xxhash_1737039922516/work
yarl @ file:///opt/miniconda3/conda-bld/yarl_1762250142216/work
zipp @ file:///opt/miniconda3/conda-bld/zipp_1763977858404/work
